01-Data-Storage-and-Batch-Pipelines/04-ETL-and-introduction-to-Airflow/02-Track-long/pyproject.toml
[tool.poetry]
name = "4-ETL-and-introduction-to-Airflow/Track-long"
version = "0.1.0"
description = ""
authors = ["S. Saunier <seb@lewagon.org>"]

[tool.poetry.dependencies]
python = "^3.8"
apache-airflow = "2.3.0"
apache-airflow-providers-postgres = "4.1.0"
lewagonde = { path = "../../../common/lewagonde/", develop = true }

[tool.poetry.dev-dependencies]
pytest = "7.1.2"
ipython = "^8.4.0"
black = "^22.3.0"
isort = "^5.10.1"
pytest-mock = "3.2.0"
apache-airflow-backport-providers-sqlite = "2021.3.3"
pytest-responses = "^0.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 160
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 160
multi_line_output = 3
include_trailing_comma = true
